+++ src/components/common/Card.tsx
@@ -1,4 +1,4 @@
-import type { KeyValuePair } from "@models/key-value-pair";
+import type { KeyValuePair } from "@models/keyValuePair";
 
 interface Props {
   title: string;
